Move date formatting to dedicated file
