+++ src/pages/projects/ProjectCard.tsx
@@ -25,37 +25,7 @@ import CategoryIcon from '../../components/CategoryIcon';
 import GitHubButton from '../../components/GitHubButton';
 import { ProjectCardListIcon } from '../../components/ProjectCardListIcon';
 import { Project } from '../../models/Project';
-
-export function formatLastUpdateAge(lastCommitAgeInDays: number | null) {
-  if (lastCommitAgeInDays == null) {
-    return 'never';
-  }
-
-  const monthDuration = 30;
-  const yearDuration = 365;
-  if (lastCommitAgeInDays < 1) {
-    return 'today';
-  }
-
-  let lastUpdateText = '';
-  if (lastCommitAgeInDays < monthDuration) {
-    lastUpdateText = `${lastCommitAgeInDays} ${
-      lastCommitAgeInDays === 1 ? 'day' : 'days'
-    }`;
-  } else if (lastCommitAgeInDays < yearDuration) {
-    const lastCommitAgeInMonths = Math.floor(lastCommitAgeInDays / monthDuration);
-    lastUpdateText = `${lastCommitAgeInMonths} ${
-      lastCommitAgeInMonths === 1 ? 'month' : 'months'
-    }`;
-  } else {
-    const lastCommitAgeInYears = Math.floor(lastCommitAgeInDays / yearDuration);
-    lastUpdateText = `${lastCommitAgeInYears} ${
-      lastCommitAgeInYears === 1 ? 'year' : 'years'
-    }`;
-  }
-
-  return lastUpdateText + ' ago';
-}
+import { formatLastUpdateAge } from '../../utils/date';
 
 function FeaturedListItem() {
   return (
