+++ src/utils/date.test.tsx
@@ -1,6 +1,6 @@
-import { formatLastUpdateAge } from './ProjectCard';
+import { formatLastUpdateAge } from './date';
 
-describe('Component: ProjectCard', () => {
+describe('utils: formatLastUpdateAge', () => {
   test.each([
     [null, 'never'],
     [0, 'today'],
@@ -9,7 +9,7 @@ describe('Component: ProjectCard', () => {
     [33, '1 month ago'],
     [67, '2 months ago'],
     [400, '1 year ago'],
-    [800, '2 years ago']
+    [800, '2 years ago'],
   ])('converts %i days readable output %s', (days, expected) => {
     expect(formatLastUpdateAge(days)).toBe(expected);
   });
